requirements.txt
streamlit==1.37.0
pillow==10.4.0
supabase==2.6.0
python-dotenv==1.0.1